# images-api/src/main/java/org/gso/images/model/Image.java
package org.gso.images.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

@Data
@Builder
@Document
@NoArgsConstructor
@AllArgsConstructor
public class Image {
    @Id
    private String id;
    private String filename;
    private byte[] data;
    private String contentType;
}